requirements.txt
# sudo apt install ffmpeg ffprobe

matplotlib==3.1.2
simplejson==3.16.0
scipy==1.4.1
numpy==1.18.1
hmmlearn==0.2.2
eyeD3==0.8.12
pydub==0.23.1
pyAudioAnalysis==0.2.5
scikit_learn==0.21.3
ffmpeg-python
websocket-client